aider: Changed table background color to transparent.
# Aider chat conversation:

USER: the background is now solid white; is it possible to just make the background see-through so i can see the background gradient? (defined in body tag "background: linear-gradient(to bottom, #f4f4f4, #e6e9ec);")

ASSISTANT: To make the table background see-through so that the gradient defined in the `body` tag is visible, we need to modify the `public\styles.css` file. We will change the background color of the table to be transparent.

Here is the *SEARCH/REPLACE* block for the necessary change:

public\styles.css
```css
<<<<<<< SEARCH
table, .styled-table, #devProgramTable, #averagesTableContainer, #liveLocalTable {
    width: auto;
    max-width: 95%;
    margin: 0 auto;
    border-collapse: collapse;
    background-color: #ffffff;
    animation: slideUpFadeIn 1.5s forwards 0.5s;
}
=======
table, .styled-table, #devProgramTable, #averagesTableContainer, #liveLocalTable {
    width: auto;
    max-width: 95%;
    margin: 0 auto;
    border-collapse: collapse;
    background-color: transparent;
    animation: slideUpFadeIn 1.5s forwards 0.5s;
}
>>>>>>> REPLACE
```